pyproject.toml
[tool.poetry]
name = "news-aggregator"
version = "0.1.0"
description = ""
authors = ["tegarsosi <tegar.sosi@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
tortoise-orm = "^0.24.0"
asyncpg = "^0.30.0"
transformers = "^4.49.0"
sentence-transformers = "^3.4.1"
newspaper3k = "^0.2.8"
celery = "^5.4.0"
redis = "^5.2.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.0"
flake8 = "^7.1.2"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
